requirements.txt
openai-whisper
sounddevice
pyperclip
black
prompt_toolkit